+++ typescript/infra/config/environments/mainnet3/warp/warpIds.ts
@@ -4,7 +4,9 @@ export enum WarpRouteIds {
   ArbitrumEthereumZircuitAMPHRETH = 'AMPHRETH/arbitrum-ethereum-zircuit',
   ArbitrumNeutronEclip = 'ECLIP/arbitrum-neutron',
   ArbitrumNeutronTIA = 'TIA/arbitrum-neutron',
+  EclipseEthereumSolanaUSDC = 'USDC/eclipsemainnet-ethereum-solanamainnet',
   EclipseEthereumSolanaUSDT = 'USDT/eclipsemainnet-ethereum-solanamainnet',
+  EclipseEthereumTETH = 'tETH/eclipsemainnet-ethereum',
   EclipseEthereumWBTC = 'WBTC/eclipsemainnet-ethereum',
   EclipseEthereumWeETHs = 'weETHs/eclipsemainnet-ethereum',
   EclipseSolanaSOL = 'SOL/eclipsemainnet-solanamainnet',
@@ -13,9 +15,6 @@ export enum WarpRouteIds {
   EclipseStrideTIA = 'TIA/eclipse-stride',
   EthereumInevmUSDC = 'USDC/ethereum-inevm',
   EthereumInevmUSDT = 'USDT/ethereum-inevm',
-  // TODO fix these
-  EthereumEclipseTETH = 'tETH/eclipsemainnet-ethereum',
-  EthereumEclipseUSDC = 'USDC/eclipsemainnet-ethereum-solanamainnet',
   EthereumSeiFastUSD = 'FASTUSD/ethereum-sei',
   EthereumVictionETH = 'ETH/ethereum-viction',
   EthereumVictionUSDC = 'USDC/ethereum-viction',
